--- src/Ast/Statement.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Serhii\GoodbyeHtml\Ast;
-
-interface Statement extends Node
-{
-}
