+++ src/Ast/Statements/BlockStatement.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Serhii\GoodbyeHtml\Ast;
+namespace Serhii\GoodbyeHtml\Ast\Statements;
 
 use Serhii\GoodbyeHtml\Token\Token;
