+++ src/Ast/Statements/Statement.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Serhii\GoodbyeHtml\Ast\Statements;
+
+use Serhii\GoodbyeHtml\Ast\Node;
+
+interface Statement extends Node
+{
+}
